pyproject.toml
[project]
name = "uml-mail"
version = "1.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.0.0",
    "anyio>=4.11.0",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
]
